+++ app/src/androidTest/java/org/feature/fox/coffee_counter/data/local/dao/ItemDaoTest.kt
@@ -8,8 +8,8 @@ import androidx.test.filters.SmallTest
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import org.feature.fox.coffee_counter.data.local.Item
 import org.feature.fox.coffee_counter.data.local.ItemDatabase
+import org.feature.fox.coffee_counter.data.local.database.tables.Item
 import org.feature.fox.coffee_counter.getOrAwaitValue
 import org.junit.After
 import org.junit.Before
