+++ app/src/main/java/org/feature/fox/coffee_counter/data/local/ItemDatabase.kt
@@ -3,6 +3,7 @@ package org.feature.fox.coffee_counter.data.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import org.feature.fox.coffee_counter.data.local.dao.ItemDao
+import org.feature.fox.coffee_counter.data.local.database.tables.Item
 
 @Database(
     entities = [Item::class],
