+++ app/src/main/java/org/feature/fox/coffee_counter/data/local/UserDatabase.kt
@@ -3,6 +3,8 @@ package org.feature.fox.coffee_counter.data.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import org.feature.fox.coffee_counter.data.local.dao.UserDao
+import org.feature.fox.coffee_counter.data.local.database.tables.Funding
+import org.feature.fox.coffee_counter.data.local.database.tables.User
 
 @Database(
     entities = [User::class, Funding::class],
